refactor: 💡 continue sales screens refactors
